Changes to Cold Ray and Cold Ring and...
Fixed bug with Lightning Rod and Helm not affecting Chain Lightning and also a bug with the damage modifier when both a +50% dmg and a "half damage" artifact are present.

Cold Ray and Cold Ring no longer affect undead creatures (including Ghosts), and Cold Ray now costs 5 mana instead of 6. It also now has a damage modifer of 18 * spellpower instead of 20 * spellpower.